+++ src/components/SamplePalette.tsx
