+++ src/components/task-board/task-sheet/TaskSheetStatus.tsx
@@ -1,8 +1,8 @@
 import { Circle } from "lucide-react";
 import React from "react";
 import LabelBadge from "./LabelBadge";
-import StatusBadge from "@/components/status-badge";
 import { useTaskStore } from "@/store/useTaskStore";
+import StatusBadge from "@/components/StatusBadge";
 
 const TaskSheetStatus = ({ taskId }: { taskId: string }) => {
   const task = useTaskStore((state) =>
